structure changes needed to make chatlog testable
