+++ src/filesys.c
@@ -1,12 +1,17 @@
 #include "filesys.h"
+#include "filesys_native.h"
 
-#include "main_native.h"
 #include "debug.h"
+#include "settings.h"
 
 #include <stdlib.h>
 
+FILE *utox_get_file(const uint8_t *name, size_t *size, UTOX_FILE_OPTS opts) {
+    return native_get_file(name, size, opts, settings.portable_mode);
+}
+
 bool utox_remove_file(const uint8_t *full_name, size_t length) {
-    return native_remove_file(full_name, length);
+    return native_remove_file(full_name, length, settings.portable_mode);
 }
 
 bool utox_move_file(const uint8_t *current_name, const uint8_t *new_name) {
