+++ src/main_native.h
@@ -84,12 +84,6 @@ uint64_t get_time(void);
 /* use the file chooser to pick an avatar and set it as the user's */
 void openfileavatar(void);
 
-void native_export_chatlog_init(uint32_t friend_number);
-
-void native_select_dir_ft(uint32_t fid, uint32_t num, FILE_TRANSFER *file);
-
-bool native_move_file(const uint8_t *current_name, const uint8_t *new_name);
-
 // Push-to-talk
 /** returns 0 if push to talk is enabled, and the button is up, else returns 1. */
 void init_ptt(void);
