+++ src/settings.h
@@ -68,6 +68,52 @@ typedef struct utox_settings {
 
 extern SETTINGS settings;
 
+/* House keeping for uTox save file. */
+#define UTOX_SAVE_VERSION 3
+typedef struct utox_save {
+    uint8_t save_version;
+    uint8_t scale;
+    uint8_t enableipv6;
+    uint8_t disableudp;
+
+    uint16_t window_x, window_y, window_width, window_height;
+    uint16_t proxy_port;
+
+    uint8_t proxyenable;
+
+    uint8_t logging_enabled : 1;
+    uint8_t audible_notifications_enabled : 1;
+    uint8_t filter : 1;
+    uint8_t audio_filtering_enabled : 1;
+    uint8_t close_to_tray : 1;
+    uint8_t start_in_tray : 1;
+    uint8_t auto_startup : 1;
+    uint8_t no_typing_notifications : 1;
+
+    uint16_t audio_device_in;
+    uint16_t audio_device_out;
+
+    uint8_t theme;
+
+    uint8_t push_to_talk         : 1;
+    uint8_t use_mini_flist       : 1;
+    uint8_t group_notifications  : 4;
+    uint8_t status_notifications : 1;
+    uint8_t zero                 : 1;
+
+    uint32_t utox_last_version; // I don't like this here either,
+    // but I'm not ready to rewrite and update this struct yet.
+
+    uint8_t auto_update         : 1;
+    uint8_t update_to_develop   : 1;
+    uint8_t send_version        : 1;
+    uint8_t zero_2              : 5;
+    uint8_t zero_3              : 8;
+
+    uint16_t unused[28];
+    uint8_t  proxy_ip[];
+} UTOX_SAVE;
+
 /*
  * Loads the config file and returns a settings struct
  */
@@ -77,5 +123,21 @@ UTOX_SAVE *config_load(void);
  * Writes save_in to the disk
  */
 void config_save(UTOX_SAVE *save_in);
+
+
+/**
+ * Saves the settings for uTox
+ *
+ * Returns a bool indicating if it succeeded or not
+ */
+bool utox_data_save_utox(UTOX_SAVE *data, size_t length);
+
+/**
+ * Loads uTox settings
+ *
+ * Returns a memory pointer of *size, the caller needs to free this
+ * Returns NULL on failure
+ */
+UTOX_SAVE *utox_data_load_utox(void);
 
 #endif
