+++ src/utox.c
@@ -3,6 +3,9 @@
 #include "commands.h"
 #include "debug.h"
 #include "dns.h"
+#include "filesys.h"
+// TODO including native.h files should never be needed, refactor filesys.h to provide necessary API
+#include "filesys_native.h"
 #include "file_transfers.h"
 #include "flist.h"
 #include "friend.h"
